refactor: move code to build_database to build_database.py
